pyproject.toml
[tool.poetry]
name = "cargo2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
wagtail = "^5.2.1"
django-allauth = "^0.58.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
